Add fixes according to comments.
